Moving utility functions to a separate file
Summary:
There is a lot going on in the instance reconciler loop
which makes adding and debugging new state machines
non-trivial. This is step one of the refactoring CLs
where we remove the unit-testable utility functions to
a separate file which helps reduce the noise for the
upcoming refactoring of the main reconciler loop.

Change-Id: Iac4a71534d1eda0a77255a2702d7390dcc9bc6f5