Support static IP address (#88)
For DR restore, it is desired to keep the load balancer IP address the same
for the old and new instances. In GCP this can be achieved by reserving
a static IP address. This cl adds an optional field loadBalancerIP to
DBNetworkServiceOptionalsGCP struct. If specified, service would be
created with the IP address provided.

Change-Id: Iee45bd6078424c7b8a9f9b74593a293e81c6350c