Reduce Oracle DB image size
The docker image build will now take place in three docker stages. In a preliminary base stage, oracle-preinstall.sh is executed to prepare the OS for an oracle DB installation. In a subsequent stage, the actual database software is installed. In the final stage, the database installation is copied to a clone of the base stage leaving behind staging and installer files to create the final DB image.

The metadata file will no longer be part of El Carro DB images because it provides redundant information which can readily be obtained from environment variables. Further, OCR images do not come with a metadata file.

Change-Id: Idabc34898713358ab4f24e21c44e970b00d38113