Create image type flag file during startup of oracledb container (#106)
There's a need to store the type of a database image during instance creation/bootstrapping. That information could be stored in the data plane or the control plane. During provisioning of a new instance, shortly after a database is bootstrapped, our current workflows do not allow us to determine whether an image was originally seeded or not because FetchServiceImageMetaData returns the name of the CDB that has just been created, sometimes leading the control plane to use the wrong bootstrapping mode and making integration tests flake. The introduction of the seeded/unseeded image type file flag allows us to confidently assert during instance provisioning whether an image was originally seeded or not. After provisioning, all images/instances are effectively considered seeded.

Change-Id: Iccba27ae8f6d504e2bdfc9519264c0304b9b3ca8