Improve detection of seeded/unseeded images
Restrict execution of the fetchMetadataFromImage() function to the oracledb container.

Relying on cdbName="" to determine whether an image is seeded or not has shown to be flaky in instances where shortly after a CDB is created and before bootstrapping, the control plane incorrectly engages the bootstrap sequence for seeded images because cdbName returned by the data plane is no longer empty. The inverse has also occurred where the control plane incorrectly engages the unseeded bootstrap sequence for seeded images. This changes forces the control plane to use whatever image type (seeded/unseeded) that was determined during startup of the oracledb container.

Change-Id: I6b140a8bae534506a5f62735eb89d0516ca1a773