Adjust log level on successful provisioning (#222)
* Adjust log level on successful provisioning

Currently a successful database provisioning operation is reported with an "Error" log level.  For example:

```
I0804 21:21:26.697425      95 common.go:193] "parent task: Done" task="Bootstrap"
E0804 21:21:26.697655      95 init_oracle.go:216] "CDB provisioning DONE"
I0804 21:21:26.699758      23 dbdaemon_proxy.go:403] "proxy/ProxyRunInitOracle: DONE"
```

This change sets the log level to the default Info verbosity, in line with the other success messages.  It shouldn't affect functionality, but rather just the levels displayed in logs.

* Remove extraneous err for info message

Now that we use InfoS for the error message, we can't (and shouldn't)
pass err as a first parameter.  As err is nil anyway, we don't need it
in this case.