Always pull Operator image
The default imagePullPolicy is IfNotPresent, images are only pulled if not already present on the kubelet. Changing the policy to Always ensures that we always pull the image if the digest of the image on the kubelet is different from the one on the image repo. This is useful for dev environments where the image same tag is reused. I've seen instances where my latest operator image did not get pulled because Kubernetes compared image tags instead of digest.

See https://kubernetes.io/docs/concepts/containers/images/#image-pull-policy for more details.

Change-Id: Ic89e4e218a97a47dec7254789f18b68908cb383b