Update rules to latest versions
This updates our bazel,compilers, runtimes, and rules to the latest
versions. New versions introduce bugs in nogo static checking, so
disable broken static checker. Stamping for container_push has changed
so we need to add the `stamp` option now.

We get some arcane protobuf dependency hell due to the gnostic
dependency in k8s. After updating it appears we must set this to legacy
protobuf mode to avoid endless conflicting proto dependencies.

With newer golang versions there is a bug introduced in godror where the
driver tries to initTZ an empty map. Lets try and leak the map and let
it get garbage collected isntead of deleting it manually. This should be
fixed correctly when we upgrade godror separately.

Bug: 243038110
Change-Id: Ic08937aeac22f9032659ca4fa7e7ebf0a1a340d8