Remove config controller
This controller is mostly unused, it attempts to reconcile instances
when a shared config change occurs but this violates the one controller
for a resource principle and makes some upcoming changes harder. Instead
let the instance controller eventually propagate changes from the
configs.

Future work would be to add a watcher on the config types from the
instance controller similar to the PITR's watch on instances.

Change-Id: Iab9c5e15e4bc01ab914fc915001d7ba8cc39a4c3