Start standardizing resource labels
We have a number of labels that we use for matching against the various
resources in El-Carro. This moves us towards a smaller set of more
flexible labels.

"instance" should be used to tag all resources associated with a
particular instance
"task-type" should be used to differentiate deployments by their
purpose, right now just "database" but our agent deployments can also be
migrated to this in the near future.

Change-Id: Ib4b60b67e43a909065533e7400db56aea5e6145c