Add custom monitoring agent
This replaces our third party monitoring agent that operated through the
DB Daemon with a custom agent that operates via regular database users.

This also adjusts reonciliation to use some of our newer best practices.
and removes a lot of code in the process.

Adds in the oracle client binaries and their license for the monitoring
pod.

Finally it increases the timeout for agent patching to match db
patching.

bug: 239772417
Change-Id: I466fa67edcdd58ad201afe62074327954d8926f7