Reconcile instances on changes to the config
This was done by the config controller previously, since we deleted it
we didnt reconcile on events from the config. That made the patching int
test somewhat more flakey so lets add back in reconcile on config events
without adding in an entire controller. This is based on the database
controller that listens to instance level events.

Also configure the instance test to not wait when deleting databases.

Change-Id: I59bed83195f704ce0ab93668ae17a9a36e0e1cf0