Fix backupschedule spec propagation
This change fixes an issue that was preventing updates to backupschedule specs from getting propagated to CronAnything specs. For some reason, updating an entire CronAnything object in operations.go after updating the status of that same CronAnything object was causing new updates to the spec to disappear. To avoid this issue, we directly update CronAnything specs from backupschedule_controller.go.

b/251830497

Change-Id: I3a5afd914fa50481ebbec4e350234aad86dadc2f