Tweaks for test flakes
Previously we would return that we do not need any further reconciling
while waiting for the instance to be ready in the instance reconciler if
the sts was complete but the database was not yet started.

Instead lets be sure to ask to be requeued a bit later.

Also log errors in removing symlinks as we see this flaking very rarely
but its unclear why the cleanup didnt run, whether it really was not
found or a different retriable error was thrown by the Lstat operation.

Change-Id: I3e01e1f08c994cd2b66d6af4aae61fc09e8f657d