Adjust timeouts in dialers
This adjusts down the fallback timeout in the dbdaemon client dialers
for network calls. In practice if the host comes up 20 or more seconds
after your dial the dial will never complete but underlying networking
may hold the connection open up to your context length. So lets adjust
down the fallback to avoid unneccessary waiting.

This was found as a cause of test flakes.

Change-Id: I9d19d05341d2d5f39117f945f45e52d9e90f5ece