Relax some timeouts on immediate reconciles
These almost always pass but sometimes a reconcile is delayed by over a
minute. Lets wait a couple minutes to give throttling a chance to
effectively delay the changes without flaking our tests.

Any further waiting is probably a sign of us under-reconciling somehow
so we should be careful before extending these further. But all of these
are based on code making specific changes to the instance.spec
immediately prior to this wait.

Change-Id: Ia82c55f0766fed41cc3a5bf718734f88ed0acc71