Stopped Instance Hang During Deletion. (#338)
In this commit we fix the hanging instance issue. If a instance is stopped and the user wishes to delete it, we will omit the logic ran that requires a running dbdaemon. This logic is only useful when deleting an individual PDB from a runnin instance, if we are deleting the instance anyways then we do not need to run SQL queries to delete the files associated with a PDB.

bug: 276327664

Change-Id: I844309f78cdf8861560049127e09302f59d9cc9e

Co-authored-by: Tabatha Lewis <tabathal@google.com>