Remove default pod AntiAffinity. (#341)
Previously, we had configured pod antiaffinty's so that pods were not scheduled on the same instance by default. We did this for historic reasons. We would like to remove this so as to make our logic more configurable and allow users to schedule multiple instances on the same node if they so desire.

bug: 277624623

Change-Id: I4b957d99ef24007a18260dcfcdb43efbc09e3021

Co-authored-by: Tabatha Lewis <tabathal@google.com>